Refactor: Split init.lua into smaller, modular files
I've broken down the main lua/maorun/time/init.lua file into several smaller, more focused modules to improve code organization and maintainability.

The new structure is as follows:

- `lua/maorun/time/config.lua`: Handles constants, default configurations, and the shared state object.
- `lua/maorun/time/utils.lua`: Contains utility functions like saving data, project/file info detection, and average calculation.
- `lua/maorun/time/core.lua`: Houses the primary business logic, including initialization, time calculations, and time manipulation functions (start, stop, add, set, etc.).
- `lua/maorun/time/autocmds.lua`: Manages the setup of all autocommands for automatic time tracking.
- `lua/maorun/time/ui.lua`: Contains the `select` function used for user interactions (inputting hours, selecting days, etc.).
- `lua/maorun/time/init.lua`: The main module file now orchestrates the other modules, provides the `setup()` function, and reconstructs the public API (`Time` table and module exports).

This refactoring aims to make the codebase easier to understand and develop further, without altering existing functionality.

Note: I encountered persistent timeouts when trying to run automated tests in the development environment. I highly recommend manual testing and review to ensure all functionality remains intact.
